+++ src/components/caution/caution.js
@@ -1,5 +1,5 @@
 import React from "react"
-import InformationBox from "./informationBox/informationBox.js"
+import InformationBox from "../informationBox/informationBox.js"
 
 export const Caution = ({ children, ...rest }) => {
   const borderColor = `rgba(218, 165, 32, 1)`
