+++ src/sections/1/1-1-high-quality-cpr.mdx
@@ -4,9 +4,9 @@ route: /sections/1/1.1-high-quality-cpr
 menu: "Section 1: General Patient Care"
 ---
 
-import EMTLevel from "../../components/emtLevel.js"
-import Note from "../../components/note.js"
-import MyLink from "../../components/myLink.js"
+import EMTLevel from "../../components/emtLevel/emtLevel.js"
+import Note from "../../components/note/note.js"
+import MyLink from "../../components/myLink/myLink.js"
 
 <Note id="note-1">
   <strong>
