+++ src/sections/2/2-1-adrenal-insufficiency-crisis.mdx
@@ -4,13 +4,11 @@ route: /sections/2/2.1-adrenal-insufficiency-crisis
 menu: "Section 2: Medical Protocols"
 ---
 
-import EMTLevel from "../../components/emtLevel.js"
-import Note from "../../components/note.js"
-import MyLink from "../../components/myLink.js"
+import EMTLevel from "../../components/emtLevel/emtLevel.js"
+import Note from "../../components/note/note.js"
+import MyLink from "../../components/myLink/myLink.js"
 import Screen from "../../utils/screen.js"
 import { css } from "@emotion/core"
-
-
 
 <EMTLevel level="basic">
   <ul>
