+++ src/sections/2/2-3A-altered-mental-neurological-status-diabetic-emergencies-coma.mdx
@@ -4,10 +4,10 @@ route: /sections/2/2.3A-altered-mental
 menu: "Section 2: Medical Protocols"
 ---
 
-import EMTLevel from "../../components/emtLevel.js"
-import Caution from "../../components/caution.js"
-import Note from "../../components/note.js"
-import MyLink from "../../components/myLink.js"
+import EMTLevel from "../../components/emtLevel/emtLevel.js"
+import Caution from "../../components/caution/caution.js"
+import Note from "../../components/note/note.js"
+import MyLink from "../../components/myLink/myLink.js"
 
 <EMTLevel level="basic">
   <ul>
